feat: add common layout for pages with navigation bars
