+++ apps/web/src/app/(app)/layout.tsx
@@ -1,0 +1,14 @@
+import Navigation from '@/components/layout/Navigation';
+
+interface AppLayoutProps {
+  children: React.ReactNode;
+}
+
+export default function AppLayout({ children }: AppLayoutProps) {
+  return (
+    <>
+      <Navigation />
+      {children}
+    </>
+  );
+}
