+++ apps/web/src/app/(app)/page.tsx
@@ -1,0 +1,3 @@
+export default async function Home() {
+  return <></>;
+}
