+++ apps/web/src/app/(app)/provider/[providerId]/page.tsx
@@ -1,12 +1,8 @@
 import { getTranslations } from 'next-intl/server';
 
+import Navigation from '@/components/layout/Navigation';
 import { Button } from '@/components/ui/button';
-import {
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-} from '@/components/ui/card';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Separator } from '@/components/ui/separator';
 
 import ReviewList from './_components/ReviewList';
@@ -85,6 +81,8 @@ export default async function ProviderPage() {
 
   return (
     <div className="min-h-screen bg-muted/30">
+      <Navigation />
+
       <div className="mx-auto flex w-full max-w-6xl flex-col gap-8 px-6 py-10">
         <section className="relative overflow-hidden rounded-3xl border bg-card shadow-sm">
           <div className="relative h-44 w-full bg-muted/60">
@@ -203,3 +201,4 @@ export default async function ProviderPage() {
     </div>
   );
 }
+
