--- apps/web/src/app/page.tsx
@@ -1,9 +1,0 @@
-import Navigation from '@/components/layout/Navigation';
-
-export default async function Home() {
-  return (
-    <>
-      <Navigation />
-    </>
-  );
-}
